Put the JavacPluginSetup functionality directly onto the Java subsystem.
