+++ src/python/pants/backend/jvm/subsystems/BUILD
@@ -27,14 +27,6 @@ python_library(
   sources = ['java.py'],
   dependencies = [
     ':zinc_language_mixin',
-    'src/python/pants/subsystem',
-  ],
-)
-
-python_library(
-  name = 'javac_plugin_setup',
-  sources = ['javac_plugin_setup.py'],
-  dependencies = [
     'src/python/pants/build_graph',
     'src/python/pants/option',
     'src/python/pants/subsystem',
