+++ src/python/pants/backend/jvm/targets/java_library.py
@@ -5,7 +5,6 @@
 from __future__ import (absolute_import, division, generators, nested_scopes, print_function,
                         unicode_literals, with_statement)
 
-from pants.backend.jvm.subsystems.javac_plugin_setup import JavacPluginSetup
 from pants.backend.jvm.targets.exportable_jvm_library import ExportableJvmLibrary
 
 
@@ -23,7 +22,7 @@ class JavaLibrary(ExportableJvmLibrary):
 
   @classmethod
   def subsystems(cls):
-    return super(JavaLibrary, cls).subsystems() + (JavacPluginSetup, )
+    return super(JavaLibrary, cls).subsystems()
 
   def __init__(self, *args, **kwargs):
     """
