+++ src/python/pants/backend/jvm/tasks/jvm_compile/BUILD
@@ -104,7 +104,6 @@ python_library(
     ':analysis_parser',
     ':analysis_tools',
     ':jvm_compile',
-    'src/python/pants/backend/jvm/subsystems:javac_plugin_setup',
     'src/python/pants/backend/jvm/subsystems:scala_platform',
     'src/python/pants/backend/jvm/subsystems:shader',
     'src/python/pants/backend/jvm/targets:java',
